refactor: 참가자 객체 생성 해당 객체에 위임
